[chore]: update watch test
